refactor : 로그인/회원가입 로직 리팩토링
