+++ src/app/api/signup/route.ts
@@ -1,12 +1,10 @@
-import type { FormState } from '@/types/auth';
 import { createClient } from '@/utils/supabase/server';
 import { NextResponse } from 'next/server';
 
 export async function POST(request: Request) {
   const supabase = createClient();
-  const { email, password, nickname } = (await request.json()) as FormState;
-  console.log(email);
-  
+  const { email, password, nickname } = await request.json();
+
   const { data, error } = await supabase.auth.signUp({
     email,
     password,
